improve sql query for tag denylist
Unfortunately, excludeTags defined in the _config.yml
cannot be used by the SQL Queries defined on a per-page
basis, since the globs are dynamic in nature, as applied
by Jekyll, so these are not picked up by our
current jekyll-sqlite implementation for generating
the query bind params.

For now, we hardcode the sql deny list, but will file
an issue on the jekyll-sqlite side to see if I ca
improve it there.