Solidify some default page attributes
Jekyll uses a default_proc to dynamically generate the page
data that comes from the _config.default block. However,
the jekyll-sqlite plugin iterates through the page data
keys to inject them as sql query parameter bindings. The plugin
cannnot use default page attributes as a result.

This commit adds a new plugin with a higher priority that loads
as soon as all posts are read, and "solidifies" a few attributes
that we know will be used this way.